[INJIMOB-3409] refactor: separet classes to files
